+++ src/fsw/index.js
@@ -8,8 +8,8 @@ import { re } from './fsw-re';
 import { parse } from './fsw-parse';
 import { compose } from './fsw-compose';
 import { info } from './fsw-info';
-import { columnDefaults, columns } from './fsw-columns';
+import { columnDefaults, columnDefaultsMerge, columns } from './fsw-columns';
 import { kind, category, group, ranges, isType } from './fsw-structure';
 import { colors, colorize } from './fsw-colorize';
 
-export { re, parse, compose, info, columnDefaults, columns, kind, category, group, ranges, isType, colors, colorize }
+export { re, parse, compose, info, columnDefaults, columnDefaultsMerge, columns, kind, category, group, ranges, isType, colors, colorize }