+++ src/swu/index.js
@@ -8,8 +8,8 @@ import { re } from './swu-re';
 import { parse, encode, decode, pair } from './swu-parse';
 import { compose } from './swu-compose';
 import { info } from './swu-info';
-import { columnDefaults, columns } from './swu-columns';
+import { columnDefaults, columnDefaultsMerge, columns } from './swu-columns';
 import { kind, category, group, ranges, isType } from './swu-structure';
 import { colors, colorize } from './swu-colorize';
 
-export { re, parse, encode, decode, pair, compose, info, columnDefaults, columns, kind, category, group, ranges, isType, colors, colorize }
+export { re, parse, encode, decode, pair, compose, info, columnDefaults, columnDefaultsMerge, columns, kind, category, group, ranges, isType, colors, colorize }