+++ src/swu/swu-columns.js
@@ -21,6 +21,52 @@ const columnDefaults = {
 }
 
 /**
+ * Function to an object of column options with default values
+ * 
+ * @function swu.columnDefaultsMerge
+ * @param {ColumnOptions} options - object of column options
+ * @returns {ColumnOptions} object of column options merged with column defaults
+ * @example
+ * swu.columnDefaultsMerge({height: 500,width:150})
+ * 
+ * return {
+ *   "height": 500,
+ *   "width": 150,
+ *   "offset": 50,
+ *   "pad": 20,
+ *   "margin": 5,
+ *   "dynamic": false,
+ *   "punctuation": {
+ *     "spacing": true,
+ *     "pad": 30,
+ *     "pull": true
+ *   },
+ *   "style": {
+ *     "detail": [
+ *       "black",
+ *       "white"
+ *     ],
+ *     "zoom": 1
+ *   }
+ * }
+ */
+const columnDefaultsMerge = (options) => {
+  if (typeof options !== 'object') options = {};
+  return {
+    ...columnDefaults,
+    ...options,
+    punctuation: {
+      ...columnDefaults.punctuation,
+      ...options.punctuation
+    },
+    style: {
+      ...columnDefaults.style,
+      ...options.style
+    }
+  }
+}
+
+/**
  * Function to transform an SWU text to an array of columns
  * 
  * @function swu.columns
@@ -101,19 +147,7 @@ const columnDefaults = {
  */
  const columns = (swuText, options) => {
   if (typeof swuText !== 'string') return {};
-  if (typeof options !== 'object') options = {};
-  const values = {
-    ...columnDefaults,
-    ...options,
-    punctuation: {
-      ...columnDefaults.punctuation,
-      ...options.punctuation
-    },
-    style: {
-      ...columnDefaults.style,
-      ...options.style
-    }
-  }
+  const values = columnDefaultsMerge(options)
 
   let input = parse.text(swuText);
   let cursor = 0;
@@ -214,4 +248,4 @@ const columnDefaults = {
   };
 }
 
-export { columnDefaults, columns }
+export { columnDefaults, columnDefaultsMerge, columns }